Improve handling of failed commands
Deprecate counter-based approach in favor of a new goroutine, which is fired after any command sent to the API. After 3 seconds, an additional check is scheduled to update the door state in HomeKit.

This emsures that a door will not remain in an Opening or Closing state forever, if there were no change in the door sensor's state (e.g. due to a failure of wiring or the physical lift)